separate visitor into stmt and expr visitor
