+++ graalphp-language/src/test/resources/org/graalphp/parser/ParseFnTest01.php
@@ -1,6 +1,7 @@
 <?php
 
 function foo($a) {
+    $a = 1;
     return 1337;
 }
