--- graalphp-language/tests/return/return-simple.output
@@ -1,5 +1,0 @@
-<?php
-
-
-
-?>
